Allow extension with custom operations
